move modeling functions to modeling_utils
